+++ js/data/game-data.js
@@ -1,3 +1,8 @@
+import getGameSingle from './../get-game-single';
+import getGameDouble from './../get-game-double';
+import getGameTriple from './../get-game-triple';
+import getStatsBlock from './../get-stats-block';
+
 const QUESTIONS_LENGTH = 10;
 
 const Answer = {
@@ -221,4 +226,21 @@ const getNextState = (state, answer) => {
   return nextState;
 };
 
-export {QUESTIONS_LENGTH, Answer, Reward, countScores, createTimer, initialState, generateQuestions, preloadImages, loadableImages, questions, GameType, getNextState};
+const getGameModule = (state) => {
+  const GameTypes = {
+    [GameType.SINGLE]: getGameSingle,
+    [GameType.DOUBLE]: getGameDouble,
+    [GameType.TRIPLE]: getGameTriple,
+  };
+  GameTypes[questions[state.level].type](state, questions[state.level]);
+};
+
+const getNextScreen = (state) => {
+  if (state.level < QUESTIONS_LENGTH && state.lives > 0) {
+    getGameModule(state);
+  } else {
+    getStatsBlock(state);
+  }
+};
+
+export {QUESTIONS_LENGTH, Answer, Reward, countScores, createTimer, initialState, generateQuestions, preloadImages, loadableImages, questions, GameType, getNextState, getGameModule, getNextScreen};
