+++ js/get-game-single.js
@@ -3,9 +3,8 @@ import renderBlock from './render-block';
 import footerTemplate from './components/footer';
 import headerTemplate from './components/header';
 import statsTemplate from './components/stats';
-import {Answer, getNextState} from './data/game-data';
+import {Answer, getNextState, getNextScreen} from './data/game-data';
 import repeatGame from './repeat-game';
-import getNextScreen from './get-next-screen';
 
 const moduleGameSingle = (state, question) => getElementFromTemplate(`${headerTemplate(state)}
   <div class="game">
