+++ js/get-stats-block.js
@@ -24,40 +24,39 @@ const moduleStats = (state) => {
           <td class="result__total"></td>
           <td class="result__total result__total--final">${resultTotal}</td>
         </tr>`;
-    } else {
-      return `<tr>
-          <td class="result__number">1.</td>
-          <td colspan="2">
-            ${statsTemplate(state)}
-          </td>
-          <td class="result__points">×&nbsp;${Reward[Answer.CORRECT]}</td>
-          <td class="result__total">${resultTotal}</td>
-        </tr>
-        <tr>
-          <td></td>
-          <td class="result__extra">Бонус за скорость:</td>
-          <td class="result__extra">${fastAnswers}&nbsp;<span class="stats__result stats__result--fast"></span></td>
-          <td class="result__points">×&nbsp;${Reward[Answer.FAST]}</td>
-          <td class="result__total">${scores[Answer.FAST]}</td>
-        </tr>
-        <tr>
-          <td></td>
-          <td class="result__extra">Бонус за жизни:</td>
-          <td class="result__extra">${lives}&nbsp;<span class="stats__result stats__result--alive"></span></td>
-          <td class="result__points">×&nbsp;${Reward.LIVE}</td>
-          <td class="result__total">${scores.LIVE}</td>
-        </tr>
-        <tr>
-          <td></td>
-          <td class="result__extra">Штраф за медлительность:</td>
-          <td class="result__extra">${slowAnswers}&nbsp;<span class="stats__result stats__result--slow"></span></td>
-          <td class="result__points">×&nbsp;${Reward[Answer.SLOW]}</td>
-          <td class="result__total">${scores[Answer.SLOW]}</td>
-        </tr>
-        <tr>
-          <td colspan="5" class="result__total result__total--final">${totalReward}</td>
-        </tr>`;
     }
+    return `<tr>
+        <td class="result__number">1.</td>
+        <td colspan="2">
+          ${statsTemplate(state)}
+        </td>
+        <td class="result__points">×&nbsp;${Reward[Answer.CORRECT]}</td>
+        <td class="result__total">${resultTotal}</td>
+      </tr>
+      <tr>
+        <td></td>
+        <td class="result__extra">Бонус за скорость:</td>
+        <td class="result__extra">${fastAnswers}&nbsp;<span class="stats__result stats__result--fast"></span></td>
+        <td class="result__points">×&nbsp;${Reward[Answer.FAST]}</td>
+        <td class="result__total">${scores[Answer.FAST]}</td>
+      </tr>
+      <tr>
+        <td></td>
+        <td class="result__extra">Бонус за жизни:</td>
+        <td class="result__extra">${lives}&nbsp;<span class="stats__result stats__result--alive"></span></td>
+        <td class="result__points">×&nbsp;${Reward.LIVE}</td>
+        <td class="result__total">${scores.LIVE}</td>
+      </tr>
+      <tr>
+        <td></td>
+        <td class="result__extra">Штраф за медлительность:</td>
+        <td class="result__extra">${slowAnswers}&nbsp;<span class="stats__result stats__result--slow"></span></td>
+        <td class="result__points">×&nbsp;${Reward[Answer.SLOW]}</td>
+        <td class="result__total">${scores[Answer.SLOW]}</td>
+      </tr>
+      <tr>
+        <td colspan="5" class="result__total result__total--final">${totalReward}</td>
+      </tr>`;
   };
 
   return getElementFromTemplate(`${headerTemplate(state)}
